# docs/mobile-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Discover Albania Transport — Mobile API
  version: "1.0.0"
  description: |
    API surface for the native Android ticket-agent app (drivers/conductors/
    agents): device activation, staff login, today's trips, ticket
    validation, ticket sales, print jobs, and offline sync. All booking,
    payment, and ticket *data* remain authoritative on this backend — the
    Android app is a client, not a second source of truth.

    Base path: `/api/mobile/v1`. Versioned from day one so a future breaking
    change doesn't break an already-installed APK in the field.

    **Production activation gate**: `tickets/sell` and real fiscal-receipt
    printing are only reachable when the login/refresh response's
    `capabilities.sellingEnabled` / `capabilities.fiscalPrintingEnabled` are
    `true` — both are hardcoded `false` in production until a real certified
    fiscalization provider is live. See the implementation plan for details.
servers:
  - url: https://transport.discoveralbania.al/api/mobile/v1
security:
  - bearerAuth: []
paths:
...
  /tickets/sell:
    post:
      summary: Sell a new ticket (cash/card-on-device)
      description: |
        Gated on the `bookings` permission and on
        `capabilities.sellingEnabled` (production-disabled until a real
        fiscalization provider is live — see the implementation plan).
        Online-only, never offline-queued. Requires an `Idempotency-Key`
        header; a retried request with the same key replays the original
        result instead of selling a second ticket.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tripDepartureId, travelDate, passengerName, passengerPhone, seats, paid]
              properties:
                tripDepartureId: { type: integer }
                travelDate: { type: string, format: date }
                passengerName: { type: string }
                passengerPhone: { type: string }
                passengerEmail: { type: string, format: email }
                seats: { type: integer, minimum: 1, maximum: 9 }
                routeStopId: { type: integer, description: Intermediate boarding stop, if not boarding at the route's start. }
                paid: { type: boolean }
                amountOverride: { type: string, description: "Overrides the computed fare+fee total, e.g. for a discount." }
      responses:
        "200":
          description: Sold
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookingId: { type: integer }
                  bookingReference: { type: string }
                  ticketToken: { type: string, nullable: true }
                  fiscalization: { $ref: "#/components/schemas/FiscalizationResult" }
        "400": { description: Missing Idempotency-Key or invalid body, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { description: Selling disabled (capabilities gate) or account lacks bookings permission, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "422": { description: Sale failed (sold out, invalid date, no price set, etc), content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Access token from /auth/login or /auth/refresh, 15-minute TTL.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
...
    FiscalizationResult:
      type: object
      properties:
        fiscalized: { type: boolean }
        nivf: { type: string }
        nslf: { type: string }
        qrData: { type: string }
        reason: { type: string }
...